+++ lbrynet/daemon/__init__.py
@@ -1,0 +1,3 @@
+from lbrynet.daemon.auth.client import LBRYAPIClient
+
+get_client = LBRYAPIClient.get_client